requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary  # PostgreSQL接続用
pydantic
pydantic[email]
pydantic-settings
alembic
ruff
bcrypt
PyJWT
python-multipart
